Make boot URL templating work in IE 11 / early Edge
Make the boot URL templating used during development work in IE 11 and
Edge. In IE 11 `document.currentScript` is not available and in IE 11 /
early Edge, the `URL` constructor is not available before polyfills have
been loaded.

The `document.currentScript` fallback here assumes that the boot script
has been loaded synchronously so that the boot script is the last
`<script>` element on the page at the point when it runs.

Also add a fast path for production builds where URLs do not contain
template params.

This is a follow up to #1441