Remove AnnotationBody `hasContent` prop
Remove the `hasContent` prop because it can be inferred by looking at
the `text` prop instead. If an annotation has been moderated, then the
`text` will either be non-empty if the user is a moderator or it will be
redacted and thus be empty if the user is not a moderator.