Make `env NODE_ENV=production gulp watch` work
Make it easier to test a production build of the client locally by
making `env NODE_ENV=production gulp watch` build code with production
JS/CSS transforms applied but still load assets from the dev server
instead of from https://cdn.hypothes.is.

Add a flag to the function that builds the boot script indicating whether we
are running the dev server, and take that into account when determining the
asset path.