Fix module load error on pages defining `require`
There are two expressions which reference a global `require` function in
the Browserify prelude. Both of them need to be rewritten to use
`hypothesisRequire` instead. Previously only the first use was
rewritten, which worked as long as the page didn't define its own
`require` function. On https://qa.hypothes.is/search for example, h's
JavaScript bundles do.