Extract `quote` function from annotation / view-filter
Both `components/annotation` and `view-filter` modules had logic and
tests for extracting a quote from an annotation object. Extract these
into a shared `quote` function in `util/annotation-metadata`.

One small change is that `view-filter`'s implementation allowed
annotation objects with _no_ `target` property. These should never
occur. There is a known issue where creating a Page Note results in an
annotation with an _empty_ `target` property until the annotation is
saved (see [1]). That case, which is still handled, looks like a mistake,
but resolving it is outside the scope of this change.

[1] https://github.com/hypothesis/client/issues/1290