Avoid effect in tag list
Simplify and optimize `TagList` a little by determining whether to
render a link directly during the render rather than using an effect.

Effects get run after the component is mounted into the DOM and painted.
In this context this means that, for first party users, the tags would
initially be rendered as spans and then a moment later, they would be
switched to links. If the links and spans were styled differently (which
I don't think they currently are), the user would see a brief flash of the
"wrong" element. Either way, it adds some more work for the browser to
do.