Prepare for transition to ES modules
In preparation for converting the client source to ES modules, in order
that the pattern library and other frontend apps all use the same module
system, change the source type for ESLint to "module" and remove the
"use strict" declarations at the top of each source file.

Babel already inserts "use strict" at the top of every module it
processes and support for transpiling ES modules is already enabled by
@babel/preset-env.

The change is only applied to modules in `src/` as build scripts
(gulpfile.js, `scripts/**.js`) are not treated as modules.