Fix a conflict between two Babel plugins
The transform-async-to-promises Babel plugin could end up inserting helper code
between the `@ngInject` comment and the associated function, breaking
the functionality of the angularjs-annotate Babel plugin. Work around
this by adding the `$inject` property to the function manually instead
of relying on the angularjs-annotate plugin to do it.

This issue only showed up locally when testing out a change to convert
the module's imports to ES module syntax.