Convert sidebar entry point to ES modules
Use the `convert-to-es-modules` script from the
hypothesis/frontend-toolkit repo to convert the sidebar entrypoint to ES
modules.

Because the code generated for `import` includes interop code to support
either ES or CommonJS modules, but migrating a module with a default
export changes its exports object, we need to migrate the module graph
"top-down" (ie. convert a module, then its dependencies).