Update annotation fixtures to better reflect annotation objects
One day we’ll `typedef` annotations and this will be better overall.
The default annotation fixture has been updated to reflect better what
data exists on a “default” annotation from the perspective of preact
components.

Updated a couple of tests that had made some assumptions based on the
old variant of this fixture.