Use custom dependency injector to instantiate services
As a step in the migration away from Angular, use the local dependency
injector implementation to replace Angular's injection system in the
creation of service objects. The constructed services are still
registered with Angular for use by Angular components which depend upon
them. Once all Angular components have been migrated, and uses of
built-in Angular services (`$rootScope`) have been replaced, this step
can be removed.

The `viewFilter` function was renamed to `ViewFilter` because it relies
on access to `this` and the new injector detects whether a function is a
class or not based on the capitalization of the name.