Decouple `wrapReactComponent` from Angular `$injector`
Pass the service container directly to `wrapReactComponent` instead of
looking up services via Angular's `$injector` service.

This removes the need to register services with Angular if only other
services or Preact components use them. This will enable us to
incrementally remove these service registrations as the migration
continues.

As a starting point, remove the "tags" service registration, as all the
UI components and services that use it have been migrated.