Make highlights "visible" to screen readers
Use CSS `::before` and `::after` to add "comment start" and "commend
end" notices around highlighted text which screen readers will read but
which are not visible.

This pattern follows the example of how highlights are presented in
Google Docs and also a suggestion in a blog post from the Paciello
Group.

This will get quite audibly "noisy" if there are a lot of overlapping
highlights, but I think it is better than not having the highlights be
visible to screen reader users at all.