Remove jQuery dependency from highlighter and add test for multiple text nodes
As part of an ongoing project to remove the annotator's jQuery
dependency, rewrite the highlighter to avoid it.

This change is also preparation for some upcoming changes to make
highlights more friendly for screen reader users by reducing the
number of `<hypothesis-highlight>` elements created for a run of text in
various situations. Since each `<hypothesis-highlight>` announces its
presence to the screen reader, this gets "noisy" if there are many such
elements.