Refactor and fix a11y in `SidebarContentError`
Refactor `SidebarContentError` to be simpler and provide a clearer
UX for the four possible states of direct-linked errors (states are based
on error type—annotation or group—and whether the user is logged in or
not).

Reuse sidebar-panel design pattern overall.

Provide an option for the user to “clear the error”, i.e., show all
annotations instead of “filtering” by the direct-linked annotation or
group. Provide a button for the user to log in if they are not currently
logged in.

Update `SearchStatusBar` to remove unneeded references to direct-linked
store information, as it will now not render if there is a direct-linked
error.