Add `lang=""` to user-generated/chosen content in sidebar UI
Add a `lang=""` attribute to the annotation quote, body and tag list in
the UI which display content chosen or generated by the user, to
indicate to the user agent that the language is unknown and may be
different than the client UI, which is currently always English.

Part of https://github.com/hypothesis/client/issues/1794.