Add accessible label to search input field
In my testing with screen readers the meaning was reasonably clear from
the fact that you had to click on a button labeled "Search annotations"
to focus the field and there is a "Search..." placeholder. Nevertheless,
this keeps automated checking tools happy.

A better solution would probably be to re-design the search UI so that there is
enough space for a visible label and information about the search
syntax, as well as enough space for the search field on smaller screens.