Assume white background by default in a11y tests
Several accessibility tests were failing incorrectly because axe-core
compared the component's foreground text color against the #ececec color
used for the body of the sidebar app, whereas the components are in fact
rendered on a white background.

Change a11y tests to assume a white background by default, as that is
the case for _most_ elements in the UI. Individual a11y tests can customize
this as appropriate.