Remove live-reload functionality from dev server
The client's dev server had functionality where it could live-update the
sidebar app when certain resources (eg. CSS bundles) were changed without
requiring a restart of the app.

None of the developers were using this functionality, and reloading the
app is not especially painful, so remove it to simplify future maintenance.