Remove non-functional `window-scroll` Angular directive
The /stream view is supposed to automatically fetch new annotations as
you scroll. However this has been broken for a very long time. The
code responsible for making this happen relied on `scroll` events on the `window`.
However the stream content has for a long time been contained in a scrollable
`.app-content-wrapper` container, so no such events were emitted.

For the moment, just remove the non-functional code as that gets us a
step closer to completing the migration away from Angular.