Add "Show" button to adder toolbar
Add a button to the adder toolbar to show the existing annotations that
relate to the selected text, along with a keyboard shortcut (the "s"
key) to trigger it. This provides a way to filter annotations in the
sidebar to those that correspond to a particular highlight or region of
text using only the keyboard.

Instead of an icon the "Show" toolbar button displays a badge whose
label is the number of annotations associated with the selection, as
that is more informative.

In implementing this, the CSS for the adder toolbar was refactored to
make the effects of several rules easier to follow and improve readability of
toolbar button labels:

 - Use of `!important` was removed. This used to be for the purpose of
   overriding any page styles which affected the adder. That issue no
   longer applies in most browsers due to the use of Shadow DOM to
   isolate the adder from the page's styles.

   This may cause a regression on some pages in older browsers which don't
   support Shadow DOM, if those pages' have generic CSS rules which
   style generic elements.

 - To make the label easier to read, the font size was slightly
   increased