Remove obsolete TODO and unnecessary effect
The adder toolbar used to set the `display` style property on its
buttons to resolve an issue with a site that would add `display:none` to
them. The linked issue no longer occurs, possibly because the use of
Shadow DOM hides the buttons from `document.querySelectorAll(...)`.

A timeout that was used to add/remove the `is-active` class from the
toolbar appears to be unnecessary. The "grow and fade in" transition
when the toolbar is shown works without it.