Remove `unloadAnnotations` method from `annotation-mapper`
This method inadvertently (?) mutated annotations in the store which
caused a crash later. Ultimately this method dispatched an event which
was handled only in one place in `root-thread.js` to call
`store.removeAnnotations()`.

This commit removes the method and associated `ANNOTATIONS_UNLOADED`
event and replaces callers with direct calls to
`store.removeAnnotations(...)`.

All uses of Angular events will soon need to be removed as part of the
migration away from Angular JS, but that will happen separately.

Fixes #1879