Add sendReadyToReceive method to fetch-config.js
This is a necessary request back to the parent app to tell it that the client’s RCP server is ready to receive incoming RCP requests. Without message, the parent APP does not know when to start sending RPC requests down from to the client. If they are sent too soon, they risk timeout.