Fix mutation violation when changing focused group
When the focused group is changed, top-level new annotations should
be moved to that newly-focused group, and the newly-focused group should
be persisted as the focused-group default for the user.

Fix mutation violation when changing the focused group: create new
`focus` method in `groups` service; use that to manage the logic of
changing to a new focused group. Remove the `GROUP_FOCUSED` event.
Manage the persisted focused-group default via the `defaults` store
module.