Add keyboard shortcuts for creating annotations and highlights
Allow the user to create an annotation using the keyboard by selecting
text (using any available method) and pressing the "a" key, or create a
highlight by pressing the "h" key.

The shortcuts are only enabled while the adder toolbar is visible. Since
the toolbar is only shown for static text selections, I think we will be
able to use single keys as the shortcuts without modifiers, and avoid
conflicts with browser and OS shortcuts.