Tune virtualized thread list debouncing
When scrolling through a long list of annotations quickly, the sidebar
could display blank space for a long period of time. This was happening
because the debounced update function waited for a period of at least
20ms with no "scroll" events, and a gap of that length could take some
time to appear if there was a queue of scroll events being delivered.

Fix the issue by adding an upper bound of 100ms after a scroll event before
recalculating the list of visible threads.

Also make sure to cancel any pending updates after the `detach` method
is called.