Remove browserify-shim
As a result of evolution of the libraries involved, this build tool is no
longer required.

A minor change was required in one test to adjust to the fact that
Angular now uses jqLite, as it does in the actual application, instead of jQuery.
This is because "jquery" is no longer automatically added to the bundle
when "angular" is added.