Replace angular-mocks in service tests
As part of the migration away from Angular, replace the use of
angular-mocks for constructing service instances in tests.

 - For tests that only have one or two dependencies, call the service
   constructor directly
 - For tests that have many dependencies, enough that the order could be
   easily mixed up if calling the constructor directly, re-use the
   `Injector` class which is used to wire up services in the real
   application

Several of these tests relied on Angular's built-in `$rootScope`
service. Use of this should go away entirely soon, but for this commit,
just replace it with a minimal mock - either an object with stubs or an
`EventEmitter`.