Add preStartServer() method to cross-origin-rcp module
When called, allows any incoming rpc request to wait in a queue and be proceeded a short time later when the actual start() method is called. This prevents potential timeouts from incoming RPC requests from LMS (or similar) that arrive after we tipped off the parent frame with a fetchConfig outgoing request but before startServer() is called.

- Rename start() to startServer()
- Use newer module export standard that matches the rest of the app