Move coverage reporting from Travis to Jenkins
Jenkins is the primary CI system and the source of truth for test
status, so it makes sense to generate coverage reports here too.

This also paves the way for potentially replacing codecov for coverage
status reporting with an alternative system in future which produces
output that devs find easier to work with.