Use `Injector` to run initialization functions
Replace Angular's `run` function which is used to run initialization
logic once all services are constructed with a new `Injector#run`
method.

This allows us to remove the dependency on Angular for running app
initialization logic which requires access to service instances. As a
result, services that are not used by remaining Angular components no
longer need to be registered with Angular at all.