Fix visibility of submenu item focus rings
Item submenus inside the `Menu` component are wrappd in a `Slider` to
enable an animated expand/collapse of the submenu. The `Slider` used
`overflow: hidden` to hide any content while collapsed or transitioning.
However this had the unintended effect of clipping focus rings which
extended beyond the bounds of the content.

Fix the issue by changing the `overflow` property to `visible` once the
slider's content is fully visible.