Add option to prevent auto-dismissal of certain toasts
For certain error messages, we do not want the notice to be dismissed
automatically as either the issue prevents further usage of the client
(eg. a problem acquiring an access token when third-party accounts are
in use) or the problem is likely to happen while the user is not
interacting with the client.

Add an `autoDismiss` option to `toastMessenger` message methods and use
it when handling access token fetching errors on pages that use
third-party accounts.