Add `hasFetchedProfile` selector
Change the representation of the un-fetched profile within the `session`
store module to make it possible to determine whether the profile has
been fetched yet or not and expose this information via a
`hasFetchedProfile` selector. The `profile` selector continues to return
a dummy "logged out" profile before the real profile is fetched. This
simplifies some of the code that uses that selector.

This makes it possible for services to distinguish between the user ID changing
because the user logged in by clicking "Log in" and completed the login
process vs. the user already being logged in from a previous session.