Add visibility hidden to iframe when closed (#2095)
Add visibility hidden to the sidebar iframe when its closed (and after it finishes animation). This prevents all interaction from the components of the sidebar.

Add an optional coerce method to the hostPageSetting so that any incoming values can be coerced if needed. This may be required for values from from via that may only be a string type.

openSidebar config value gets coerced to a boolean in the annotator settings (just like it does in the host-config).