Focus first item in menu when opened by any input method
Previously the `Menu` component would focus the first item in the list if it was
opened by pressing Enter or Space but not if it was clicked. This meant that it
was not possible to mix input modalities (open menu with mouse, navigate using
keyboard) easily when using the menu.  More importantly if the menu was opened
using a screen reader shortcut (eg. VoiceOver Key + Space on macOS) per the
screen reader's verbal instructions, the focus didn't change. My understanding
is that SR shortcuts generally simulate mouse rather than keyboard input.